Solve issues with code
