+++ app/src/main/java/com/example/androidapp/ui/home/HomeFragment.java
@@ -14,10 +14,12 @@ import androidx.lifecycle.ViewModelProvider;
 
 import com.example.androidapp.R;
 import com.example.androidapp.databinding.FragmentHomeBinding;
+import com.example.androidapp.sensors.SensorController;
 
 public class HomeFragment extends Fragment {
 
     private FragmentHomeBinding binding;
+    private SensorController sensorController;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
@@ -27,6 +29,8 @@ public class HomeFragment extends Fragment {
         View root = binding.getRoot();
         final TextView textView = binding.textHome;
         textView.setText("this is Home fragment");
+
+        this.sensorController = new SensorController(getContext());
 
         return root;
     }
