+++ app/src/main/java/com/example/androidapp/ui/tasks/TasksFragment.java
@@ -28,7 +28,7 @@ public class TasksFragment extends Fragment implements AdapterView.OnItemClickLi
 
 
     private FragmentTasksBinding binding;
-    List<Task> list_of_tasks;
+    List<Task> tasksList;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
@@ -44,9 +44,9 @@ public class TasksFragment extends Fragment implements AdapterView.OnItemClickLi
 
         TaskController tc = new TaskController(root.getContext());
 
-        list_of_tasks = databaseHelper.loadAvailableTasks(root.getContext());
+        tasksList = databaseHelper.loadAvailableTasks(root.getContext());
         ArrayList<Task> array_list_tasks=new ArrayList<>();
-        array_list_tasks.addAll(list_of_tasks);
+        array_list_tasks.addAll(tasksList);
         TasksAdapter adapter = new TasksAdapter(getActivity(),array_list_tasks);
         myListView.setAdapter(adapter);
 
@@ -62,7 +62,7 @@ public class TasksFragment extends Fragment implements AdapterView.OnItemClickLi
     @Override
     public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
        Log.i("TaskFragment", "You clicked Item: " + id + " at position:" + position);
-       Task selected_task= list_of_tasks.get(position);
+       Task selected_task= tasksList.get(position);
         if(selected_task.isCompleted()==false)
             Toast.makeText(getActivity(),"This task still need to be completed",Toast.LENGTH_SHORT).show();
         else
